Changes to be committed: 	modified:   etc/linux_init/clean.sh 	modified:   etc/linux_init/symlink.sh 	modified:   etc/windows_init/symlink.bat

+++ etc/linux_init/clean.sh
@@ -8,4 +8,6 @@ rm -rf ~/.vimrc
 rm -rf ~/.viminfo
 rm -rf ~/.config/nvim
 rm -rf ~/.gitignore
-sudo rm -rf /etc/wsl.conf
+if [ -n "$(which wslpath)" ]; then
+  sudo rm -rf /etc/wsl.conf
+fi
